[ARM] Add barebones support for arm.
Base the run of coremark on gdb through usb debug communication rather then UART.
The ee_printf function is implemented as a wrtie to a specific memory section where
GDB can read the essential information to extract the score.
Add a c program that extracts the information from the gdb output in run2.log.
The real size of the executable can be calculate with the size command.

extract_score requires the MHz of the target CPU as argument to run.

Add .gitignore