[MIPS] Add initial port of barebones mips.
This should work through mdb in a similar way to arm.
The initial command file for mdb is added with the name of the device.
The next steps will be to find a suitable linker script and if necessary
a startup file, and to add ee_print code that can be used by mdb to examine
the content of cycle counter registers or of a reserved memory location.
Similarly to arm.

Another possible approach will be to use UART.