[ARM] Add script that wraps a whole execution of coremark for ARM.
The script can be configured to accept a path to the desired
tool chain and cc and ld options.
It performs a clean build, a performance run (which the calculation
of the scores will be based upon), another clean build and a
validation run.
It will eventually call extract_score.sh.